design: 프로젝트 둘러보기 버튼 상단으로 이동
- 유저 접근성을 위해 프로젝트 둘러보기 버튼 하단에서 상단으로 이동
- 상단에 해당 기능이 있기 때문에 하단의 버튼을 삭제하고 둘러보기 프로젝트 개수 추가(5 -> 10)
