+++ src/features/project/components/latest-projects/LatestProjects.tsx
@@ -2,7 +2,7 @@ import { getLatestProjects } from '../../services';
 import { ProjectGallery } from '../project-gallery';
 
 export async function LatestProjects() {
-  const { data: projects } = await getLatestProjects(null, 0, 5);
+  const { data: projects } = await getLatestProjects(null, 0);
 
   return <ProjectGallery projects={projects} />;
 }
