+++ src/features/project/components/ranked-projects/RankedProjects.tsx
@@ -1,10 +1,10 @@
 import { getRankedProjects, getSnapshot } from '../../services';
-import { SimpleCardList } from '../simple-card-list';
+import { SimpleProjects } from '../simple-projects';
 
 export async function RankedProjects() {
   const { id } = await getSnapshot();
 
   const { data: projects } = await getRankedProjects(id, 0, 3);
 
-  return <SimpleCardList projects={projects} />;
+  return <SimpleProjects projects={projects} />;
 }
