+++ src/features/project/components/simple-projects/index.ts
@@ -1,0 +1,1 @@
+export * from './SimpleProjects';
